refactor: 상품 목록 조회 로직 useGetProductsQuery 훅으로 분리
