+++ src/hooks/useGetProductQuery.ts
@@ -1,0 +1,23 @@
+import useData from "./useData";
+import { ProductItemType } from "@/types/product";
+import { getProducts } from "@/apis/products/getProducts";
+import { useCallback } from "react";
+import {
+  FilterOption,
+  SortOption,
+} from "@/components/Product/Content/ProductContent.type";
+
+const useGeyProductQuery = (
+  filterOption: FilterOption,
+  sortOption: SortOption
+) => {
+  const fetchProducts = useCallback(() => {
+    return getProducts({ filterOption, sortOption });
+  }, [filterOption, sortOption]);
+  return useData<ProductItemType[]>({
+    fetchFn: fetchProducts,
+    name: "productData",
+  });
+};
+
+export default useGeyProductQuery;
